my-appointments: show list appointments in ascending time order, not descending
The Medicover API returns the user's appointments in decreasing order of time. However
when displaying the results to the user, it would be more readable to display the most
recent appointments at the bottom of the screen. Currently, when the results are large,
the oldest results are visible and scrolling is required to see more recent results.
This behavior is now changed by appending each retrieved appointment to the beginning
of the results list instead of the end.