my-appointments: optimize by moving time filtering to MedicoverSession method
Instead of fetching all of the user's appointments and then, if only future ones are
to be shown, filtering them by appointment time, combine filtering with fetching.
This allows to fetch only those pages from the API that contain future appoitnemtns,
stopping the loop when the first appointment in the past is found.
The resulting code runs faster and makes fewer calls to the Medicover API,
when my-appointments command is run without --show-past option.